sd: sync to master-558-8afbeb6
