+++ otherarch/sdcpp/tokenize_util.cpp
@@ -1,4 +1,4 @@
-﻿#include <algorithm>
+#include <algorithm>
 #include <iostream>
 #include <string>
 #include <vector>
